clean up, rename, add comments
